edit user_exercise schema to change progress and goal from integers to floats

+++ db/migrate/20180710190956_add_progress_made_and_goal_and_goal_category_to_user_exercises.rb
@@ -1,7 +1,7 @@
 class AddProgressMadeAndGoalAndGoalCategoryToUserExercises < ActiveRecord::Migration[5.2]
   def change
-    add_column :user_exercises, :progress, :integer
-    add_column :user_exercises, :goal, :integer
+    add_column :user_exercises, :progress, :float
+    add_column :user_exercises, :goal, :float
     add_column :user_exercises, :goal_category, :string
   end
 end
